Load the plugins before creating the database and ignore error when insert the status
Loading the plugins allows adding their dedicated table to the database
when calling create_tables.
Next to this, we do not really care if inserting the defaults status fails
because that will likely not occur if we are creating the DB from nothing
while it is annoying if it occurs while we are just re-running createdb
to add the table for a new plugin.