Update the format_loc filter
With the new format_loc filter if a commit is provided at the same time
as the line of code, the filter will add an extra column (hidden by
default) containing the image that the user should click on to add a
comment on that line of code.
This new filter will also add some meta information used by our tool to
find out which row of which commit the comment should be added to.