Add support in the UI for uploading files to an issue
This commit adds the Javascript allowing drag&drop mechanism to upload
a file to a ticket.
As an alternative, we also offer in the UI the possibility to use the
more traditional `upload` button.
All the upload are preformed directly on the fly with a progress bar
showing up.

Fixes http://209.132.184.222/progit/issue/52