If request access the update endpoint using GET, redirects them to the issue
If the user's session times-out and the user tries to add a comment to
an issue, they will be redirected to the login page which in turn will
redirect them to the update endpoint which can only work with POST
queries.
So quickly redirect the user to the issue page without doing anything
else.