Clone all the remote branches when cloning a project
Currently, cloning a project meant only cloning the master branch.
With this change, all the branches present in the main project will
be cloned in the fork as well.

Note: we could use the --mirror recipe from:
https://github.com/libgit2/pygit2/blob/master/docs/recipes/git-clone-mirror.rst
But this works only for pygit2 0.22+ while we want to work with 0.21 as
well.