Make the PullRequest table ready for remote pull-request
Remote pull-request are pull-request coming from a git repo that is not
local to pagure (for example hosted on gitlab).
So a PullRequest object may not have a project_id_from but may have
instead a remote_git address.
Anyway, either project_id_from or remote_git must be defined.