Fix authorized_keys file creation
Explanation is in the comments in the commit, but a short recap:
gitolite when creating an authorized_keys file, will ignore any keyfiles
that have more then a single line in them:
"WARNING: keydir/<username>.pub does not contain exactly 1 line; ignoring".

So instead if relying only on gitolite to do this, we do it manually, since
we do want people to be able to add multiple keys to their account.
We do however also update the files in the keydir used by gitolite, so that
if for some reason our regeneration fails, we at least have the first keys
for users installed, which will likely be enough for the majority of users.

Signed-off-by: Patrick Uiterwijk <puiterwijk@redhat.com>