Only hides sections if the user is logged in
Otherwise, we may end-up with an empty front-page if we hide all the
projects, in that case if the user is logged out the toggle won't be
shown and all the projects will be hidden (since we cannot show his/her
projects/forks). This way, we make sure to show all the projects in the
front page is the user is logged out but hide them by default is the
user is logged in.