Fix the user info page when we hide some section on the front page
For running pagure as front-end for pkgs.fp.o we added the option to
hide by default some sections (such as all the repos) when the user is
logged out.
This lead to the situation where we hide all the projects on the front
page but this hides the user's projects on the user's page.

With this page, we can turn off the hide feature for the user's page.