Make the SHOW_PROJECTS_INDEX list what should be showing in the index page
Until now despite its name, we were using this configuration key as a way
to hide sections from the index page.
With this change, we actually use it to list what should show on the index
page. This makes it easier to write the proper if statements to show the
sections and the corresponding switch buttons.