Handle the situation where the author of the commit did not specify an email
We have the situation in one of our repo that the author of a commit
specified his name but not his email. This leads to search_user returning
the list of all the users in the DB instead of just the one we want and
this makes pagure crash (crash that's been reported to us by email).
With this fix, is the author has no email, we'll just return the provided
name without searching further