started changes to the index_auth page
First round of changes for the index_auth / user dashboard page
Trying to make it more useful for a user, with icons showing
the issues and PRs open for each of their repos.

This also has less information about the repos presented
by default to the user (ntoably the repo description), as
the assumption here is that the user probably already knows
what the repo does, because they created it.