added more information to the commit list page
now shows the project and branch name at the top of the content
area, so the user knows what branch they are looking at for sure.

also, provides a alert box at the top if the branch being browsed
has commits ahead of the main branch / upstream project. the alert
box also has a collapsed area that shows just the commits that are
different from the upstream / main branch.