fix for #682: made the PR actions a buttonbar
Made the PR actions at the top of the PR into a buttonbar
to make it look similar to the other buttons in for eding
comments. (but left them as coloured buttons)

I had to move the form call outside of the button-group
class as it was causing issues with the styling.