Add a way to show a contributing information block when opening a new PR
When someone opens a new PR, we'll look into the requests git repo for
a file ``templates/contributing.md`` and if present, we will show it on
the new PR page to the user.

This will allow developers to express what they would like to see in a
PR when submitted.

fixes https://pagure.io/pagure/issue/587