Prevent the infinite loop of log-out/log-in
If an user logs out, they are redirected to the page they were at,
if that page requires log in, they are redirected to the login page
and if they don't pay attention, they end-up at the point they started
and logged in.

This breaks this loop by recording when someone just logged out and
breaking the redirect to the login page in this situation.