Fix checking if the user is a repo admin or not
In the previous state we were re-defining the user object in the
inner loop when checking if the user is one of the admins of the
project.
This means that when we reached the last check (is the user in a
group having commits) the user we were looking was no longer the
one we started with.

Thanks @pbrobinson for reporting the bug