fill initial comment of the create PR form if only one commit
When creating a Pull request with only 1 commit in it, the Pull
Request title is automatically filled out with the title of the
only commit.

Now, in the case of having one commit message, the body text
of the commit is now automatically filled out in the new pull
request description field (inital comment field) too.