Drop the dedicated fork folder
Instead of having a folder dedicated to forks, just place them in a
`forks` folder under the main folders (repos, tickets, requests, docs...).

This should help reducing the risk of conflicts with for example a
project called `pingou/test` when there is already a fork of the project
`test` by `pingou`.