make bottom pagination links on project list respect the sorting filter
Previously, if a sorting filter was enabled, and the pagination
links at the bottom of the Browse Projects page were used,
the sort would no longer apply. It did, however work if the
top pagination links were used. With this commit, the pagination
links at the top and bottom both respect the current sorting filter

Fixes: #1013