Do not render the diffs of large added and removed files
Rendering large diffs in PRs is really slow. To counter this
this commit does not render diffs for added and removed files
that are over 100 lines, instead it provides a link to view the file.

In the case of a removed file, it provides a link to the file as it
was the before the commit.