Update parts of the repo settings to new UI
This updates a few leftover elements that were missed
when implementing the new UI.

The pages that let you add groups and users to a repo
are now styled up with fedora-bootstrap, and use the
same user search UI as adding a user to a ticket.

The buttons to accept changes to Hooks settings are now
styled up in the same style as all the other buttons.

Fixes #908