Split out the commits page to a template of its own
Previously, the commits page, and the overview page
were mashed together in a single template. This was
a holdover from the old UI where they shared a lot
more in common. Now they do not share anything in
common, so to be in line with all the other templates,
this splits the commits into it's own template.