Fix diffs not showing for additions under 1000 lines
Commit
https://pagure.io/pagure/c/b771bf3f4fa73e1f1544114bf799e3a2361feafa
Fixed the issues for the multiple versions of pygit2 that we support,
however it also introduced a bug where additions with < 1000 lines
would not show anything (either a diff or a message saying too large).

Turns out we previously set a jinja variable that stored the linesadded
amount (with support for multiple versions of libgit2), so this commit
just uses that variable.