Fix the no readme box showing on empty repos
We recently added the a box on a projects info
page to show if there was no readme in a non-empty
repo. However, this was showing up on emtpy repos
too, in addtion to the box saying the repo has not been
inited. This update removes the noreadme box from showing
on empty repos.