Convert filenames with utf8 chars when outputting them in templates
If a git repo has a filename in it with a utf8 encoded char,
the template would previously fail to render with an error like:

    UnicodeDecodeError: 'ascii' codec can't decode byte 0xc3 in position 1:
ordinal not in range(128)

This adds a filter to convert the utf8 charaters in the templates.

The tests have also been updated to add a file to one of the repos
with a non-ascii char.