remove the branch chooser from the repoheader
Previously, there was a branch chooser dropdown that was in
the repoheader, up near the fork button. This is now removed
from there, and placed at the top of the pages where this applies.

This dropdown would only show on certian pages in a repo, and could
lead to confusion as to why it wasnt showing, so now on the files pages,
and the commits pages, it is at the top right of the content of those
tabs.

On the overview page, we always had the full branch listing there,
so there is no dropdown.

Also as part of this change, i restyled teh fork button.