Add test for issues to private repo and changes to fix those test
The test ensures that issues made to private repo are viewed by only
the owner or added user to the repo.

Changes in 'ui/issues.py' are made to put a check that new issues can
only be created by the authorize owner and all CRUD operation on issues
to a private repo can and should only be carried out by the authorize
owner of the repo.