Add check for api endpoint for private repo
The check in api endpoint to fetch tags of a repo, a condition is
introducet keeping the nature of the request as 'GET' and test case is
written accordingly.

Basically to fetch tag of a private user user needs to be logged in. If
not then the is_repo_condition will fail and no projects will be
returned.

A new error code needs to be introduced since now we have projecsts
which can or cannot be accessible to the user.