Add a different method to fetch project
`get_authorized_project()` is introduced in `pagure/__init__.py` to get
project and filter it in a way to only return project which the logged
in user is authorized to view.

Previous approach of passing a function to the as a parameter may fail
in some aspect hence it is avoided.