Fix unaccessible fork after the main repo is deleted
A new column `_is_fork` is introduced into the `projects` which keeps a
track on the repo being `fork` or `not`. This changes the function
`is_fork` and `fullname` by altering the condition. This is done to
maintain the `integrity` of the FK parent_id. The changes in UI is to
check if the main repo is deleted how the UI should function.

Also if `repo.is_fork` is true and parent doesn't exist this was
throwing error while mapping repo path hence the check for `repo.parent`
is included.

Alembic script is included for the migration of the database.