Fix test for private-repo and rename get_project to _get_project
The renaming is done on puiterwijk concern for open-points for
get_project. This is done so any function or component using it will
stop function as it should use get_authorized_project instead.

The test for private-repo still has _get_project because getting a user
authenticated demands to have an active request which is not possible
while testing the component.