Made the interaction with the watch button clearer
Previously, it was unclear to the user if they were adding
a watch or removing a watch when interacting with the watch
dropdown.

This commit changes this so there is only ever one item
showing in the watch dropdown, so it is clear to the user
what will happen when pressing that button. This also simplifies
the JS required to get this working.

There is also a style change that goes along with this, making the
buttons fit nicer with the newer, smaller fork button.

I also had to move the fork button form a bit to get it working better
with the button group styles.