Fix docs and toggle of Pagure CI plugin, add pagure_ci_token
Add proper ussage of Pagure CI in docs. Previously there was no way of
deactivating the plugin, with the addition of condition and logging the
error.

pagure_ci_token is a way of authentication added to make sure that the
POST request is made from Jenkins only and not from any external source.