Add padding to markup view and add readme display to treeview
Add a bit of padding around  markup type files when displaying
them in the file view.

Also, when viewing a directory in the tree view, if that directory
has a README in it, show the readme under the file listing on the
treeview page. This will work for both the root dir and any other
directory.