When adding files to the a git repo, rely on the user object from the db
We need this change in order to get the `local` auth working properly.
Otherwise, pagure.lib.git.add_file_to_git relied on the given user object
having a `.email` which when using the `local` auth is not true and
therefore leads to a crash.
With this change, we retrieve the user object from the database and
therefore we can just use the default email set to the user there.