Several improvements to user freedom on JavaScript land
This commit basically excludes all minified JavaScript "source code"
from pagure, and replaces them by the corresponding, true source code
provided by the upstream projects.  This has the obvious and enormous
benefit of actually respecting the Freedom that the user of pagure has
of studying the code that comes with the software, but it also has the
very important side effect of giving the user a chance to guarantee
that the JavaScript code ran by pagure is actually the one that came
from upstream.

The idea for this came from the Debian packaging of pagure, which has
to deal with the issues caused by the several minified JS code
bundled.  Initially, I was just going to replace a few JS libraries,
but I decided to go ahead and fix everything.  I have also reorganized
the structure of files a little bit, so that now we record (on the
file name) the version of the JS library we're using, but, with
symlinks, we do not depend on that version when loading the library
from the HTML pages.

Lastly, I went ahead and updated a few libraries with their latest
upstream release.  Pagure was already using pretty recent code, so the
changes introduced by the updates should not be big.