Fix displaying the branches available for a new PR in the requests template
We have changed the format of the data returned by the endpoint listing
the branches ready to be in a PR so that it also includes branches that
are already in a PR (the info is then displayed on the front page of the
project).
However, this same internal API endpoint is also called in the page
listing all the pull-requests and this page had not been adjusted for
this data format change.

This commit fixes that.

Fixes https://pagure.io/pagure/issue/1310