Fix figuring out if the milestone of a ticket was edited or not
Basically, before we always added an empty choice to the list of
milestones. This lead to a situation where there was always something
assigned as milestone even when the project had none.
With this change, we first check if there is a milestone before
adding the empty choice to the list of choices and before retrieving
the milestone from the form.