Note that SEND_EMAIL=False does not disable EMAIL_ERROR
I discovered today that when I set ``SEND_EMAIL = False`` in the Pagure
configuration, unhandled exceptions were still emailed to pingou which
is embarassing for me and noise for him. This adds a note to the
configuration documentation so that those who come after me do not make
the same mistake.

Signed-off-by: Jeremy Cline <jeremy@jcline.org>