Add a link to the user on the commit list
This adds a link to the user on the commit list.
It required a little more playing than just adding
the link, as the previous iteration had a single tag
that just wrapped everything. So we had to pull this out,
and add a link to the user on the username text.