Adjust the markdown processor to have 1 regex for all cross-project links
Until now we had two regex/extensions for the cross-project links. One
for forks one for regular project but this lead to the situation where
if you have a pattern such as <user>/<project>#<id> the first part
(<user>) would be linked to the fork of the project <project> of that
user and the second part (<project>#<id>) would link to the ticket/PR
with this id in the original project.
That was because the regex for cross-link project was matching the
content of the link generated by the cross-link fork.
By moving to one regex for all, we fix this situation.