Fix rendering raw file when the sha1 provided is one of a blob
In the regular UI, you can view the content of a file at any sha1 provided
be it one of a commit (the expected/default being the sha1 of a commit).
For binary files, the UI would then point you to the raw endpoint.
But if you were trying to see a binary file in a blob, the raw endpoint
would just error out since it was considering the commit object to be a
commit not a blob and thus ``commit.tree`` would fail.

With this change, pagure will happily return you whatever is in this blob
instead of failing.