If the form is not validating and the user is using js return that it's not ok
This should help debugging issues with the update_issue endpoint as
at the moment it would return a 200 ok when things didn't work (for
example the form did not validate).
Now instead we will return a 200 notok (that the JS client knows how to
handle) and some information about what went wrong.