Handle namespaced projects in SSE server
The SSE server's code for parsing the URL to get to the project
does not handle namespacing. With this, it does. I tried to
figure out the 'rules' for the path as best I could from other
code in the project.

Also refactor a bit to split out a path parsing function and
separate object finding functions, and be more explicit about
the specific object types we support.

Also add a test suite for the streaming server, and update all
docs, Ansible play, spec file etc. for the rename from
'pagure-stream-server' to 'pagure_stream_server'.

Fixes https://pagure.io/pagure/issue/1532

Signed-off-by: Adam Williamson <awilliam@redhat.com>