tests: do local pylint disabling properly
This is the right way to disable pylint messages for a single
line. It's not right to put the disable comment on its own
line directly above the line you want to disable the message
for. A pylint disable comment on its own line disables the
message in that *block level* - so if the comment is outside
of any class or function, it disables the message for the
entire rest of the file. See:
https://pylint.readthedocs.io/en/latest/faq.html#message-control

Merges https://pagure.io/pagure/pull-request/1568