Drop UPLOAD_FOLDER in favor of UPLOAD_FOLDER_URL
This allows specifying in the configuration file the full URL to where
the releases are available, which allows serving them in a different
domain than the main application, which is most interesting for security
reasons since there are no check on the content uploaded.

We also adjust the documentation about configuration for this purpose.