Include regular contributors when checking if user is watching a project
We were checking if the user had set a specific 'watch' flag, if they
were the creator of the project or part of a group of contributor but
we forgot to check if the user was a regular contributor who got commit
to the project.

This fixes it, with unit-tests.

Fixes https://pagure.io/pagure/issue/1601