Issue 1597- RFE - add a file attachment section to Issues page
It is common for many files/patches/images, etc, to be attached to a
single Issue. Having a section listing all the attachments is a nice
convenience, especially for issues with many comments.

The attachment section will list all the attachments(as links).  The
attachment row will include a link to the file, the date it was added,
and a link to the Comment that the file was attached to.  If there are
no attachments then do not display the "attachments section"

https://pagure.io/pagure/issue/1597