Fix the git blame view when the committer has no name
In the git tree of the linux kernel, the commit `febf7ea4bedcd36fba0843db726bba28d22bf89a`
has for author ``Author:  <sam@mars.ravnborg.org>`` which, as you can see,
has no name specified.
This makes pygit2 un-happy and basically makes it crash.

In this commit, we work around the issue by not specifying an user if
the committer did not use one.