Adjust the wording when showing a merge commit
Instead of saying that the merge was authored and committed by the
person who did the merge, we now say that the merge was `Merged` and
committed by the person who did the merge.
It is a little redundant, but since we include the diff of the merge on
the page some people find it confusing as it could be interpreted as the
person who merged did all the work (while they only did the merge).

Fixes https://pagure.io/pagure/issue/1650