Small code cleaning and moving around
We had a bunch of ``if repo_admin`` one after the other, that wasn't
quite needed, one is enough.

We moved updating the status near the top since that's one action that
is allowed for admins and the person who created the ticket. Everything
else will only be allowed for admins.

While this piece of code could still use a nice refactoring, this already
helps a little on readability and logic.