Add the logcom service
This service aims at logging the commits to the pagure_logs table outside
of the push process (ie outside of git push).
This will save time as otherwise we were logging all the commits in the
git push, so if you were pushing, for example, the linux kernel tree, it
was taking up to 45 minutes on a local push (from within the same machine).
With this change, it will still take that long to log all the commits and
save them to the DB, but at least, we will end the git push sooner.