Introduce the USER_NAMESPACE configuration key
When this configuration key is set to True, the default namespace of the
project is the user's username, basically as does github or gitlab.

This isn't really how we envisioned pagure at the beginning but this is
a request that has came up a few times and which this configuration key
easily allows now that we have full namespaces.

So basically, this commit fixes https://pagure.io/pagure/issue/1231