If no assignee is provided, do not try updating the assignee field
The method pagure.lib.add_issue_assignee() always returns a message,
so if we call this method while there are no assignee provided, it
returns a 'Nothing to change' message that we display to our user,
while in fact things did change, just not the assignee.
With this change, we no longer have this un-justified message.