Issue 1711 - Improve Issue roadmap page
The roapmap page appears to have been started, but not finished.  Certain features did not work, like sorting on tags (tags were previosuly being treated as milestones).

This fix now makes it possible to isolate individual milestones, status, and tags.  All three can be combined to create a new report.  Each criteria is clicked to select and deselect.

The layout of the page has also been improved to be more consistent, and uniform

https://pagure.io/pagure/issue/1711