Fix logging who did the action
The issue was that we relied on the object, so for comments we were
relying on the ticket or pull-request. So instead of logging that the
user `foo` commented on a ticket opened by the user `bar`, we were
logging that the user `bar` commented.

Fixes https://pagure.io/pagure/issue/1731