Improve the error message on error pulling changes of remote pull-request
With remote pull-request, the changes are pulled from the remote repo
when the PR is accessed, in order to offer to the reviewer, the latest
changes.
If for some reason the pull errors out, it will raise a GitError.
With this PR we catch this exception and return it to the user with an
error message, being a little more friendly than we were.

Fixes https://pagure.io/pagure/issue/1755