Add custom drop down lists to custom fields
This patch adds an option to create a "List" as a custom field.  To do this we need to add "key_data" to the issueKey model for storing the list items.Currently this is only used by "lists", so other field types ignore and remove the data if it's set.

https://pagure.io/pagure/issue/1748