Rework a little the active milestone/all milestone filtering
So we want either only the active milestones, or all of them, so this
is a simple boolean, if we want only the active milestones then we
retrieve the list of all the active milestones (ie: having at least one
ticket open) and consider only this list.
Otherwise, we get the list of all the milestones and consider this list.