Fix triggering web-hook notifications via fedmsg upon commit
If user turns on the fedmsg hook, we will notify on fedmsg about new
commits. The idea being that if the project has web-hooks activated as
well, we should also send web-hook notifications.
Except that since the redis connection wasn't set, we were not doing it.
With this commit, we fix that by passing the redis connection around.