Add the EXCLUDE_GROUP_INDEX feature and corresponding tests
This new configuration key allows to hide projects an user has only
access to via one or more specified group. This will not remove the
access, just not show the project on the user's user page in order
to not show the 17,000 projects that the people in the provenpackager
group in Fedora have access to when pagure is used as a front-end
for Fedora's git repositories.