Add a way to view patch attached to a ticket as raw
We added support to present nicely patch attached to a ticket but that
was without taking into account the use-case of downloading it.
With this commit there is now a ?raw=True argument that can be used
to by-pass the logic making the patch pretty and allowing to download it.