Fix support for namespaced projects in the doc server
We have added support for namespaced projects in pagure and we need to
support hosting documentation for them but we have been using the
following route:
@APP.route('/<repo>/<path:filename>')
This means that we are unable to distinguish between:
/repo/folder/file
and
/namespace/repo/file

This commit fixes that making the URLs be namespace.repo before listing
the folder and files.

This commit also adjusts the unit-tests accordingly.