Apply EXCLUDE_GROUP_INDEX to the watchlist as well
The watchlist contains all the project the user is watching as well as
all the project they have access to. So we need to apply EXCLUDE_GROUP_INDEX
to this list as well as otherwise we would end up in the same situation
where we would list all the projects there is on that pagure instance
for some users.