Move usage.rst to be the index of usage/
It was confusing to have both a usage.rst and a usage directory, so this
just moves the usage.rst to be the index of the usage directory. This
also moves the usage guide to the top of the TOC after the overview
since most people will be looking for the user guide.

Signed-off-by: Jeremy Cline <jeremy@jcline.org>