Ignore the ``No Content-Type`` error raised by pygit2
We recently starting running into this error for remote PR opened from a
github git repo. Maybe something change in github that is now triggering
this.
Anyway, with this commit, we will now be ignoring that specific error
and keep on going as usual.
It will at least allow accessing pull-requests that have already been
open and which are basically still valid.

Fixes https://pagure.io/pagure/issue/1863