Fix deleting tickets that have a tag attached
We had a bug where we could not delete issues/tickets that had been
tagged because of a lack of propagation of the change at the DB level.

With this commit we can now safely delete issues/tickets that have been
tagged and we have unit-tests to prevent this from happening again.