Add a feature to ignore existing git repo on disk
In the traditional use of pagure, if a repo exists on disk we want to not
create the project, but when using pagure as a front-end for a system
that handles creating the project on disk via a different process, in that
situation we want to be able to create the project in the DB while ignoring
the fact that there is/are already git repos on disk.

The flag ``ignore_existing_repo`` added to pagure.lib.new_project() will
allow that.

Add corresponding tests