Add functionality for dealing with issues with no milestone
We need a way find issues that do not have a milestone set yet, and we need a way to find tickets that have specific milestones as well as no milestone.  For example:

When new tickets are created they don't have a milestone, but we also have a NEEDS_TRIAGE milestone.  So to find all the issues that need to triaged(aka have a milestone set) we need to find both types.

This patch adds a new "No milestone" button to the roadmap page, and it allows you to search issues by using multiple milestones(including "none"):

http://localhost.localdomain:5000/TEST_PROJ/issues?milestone=none&milestone=0.0

We can save this as a custom report for finding all the issues that need to be triaged.