Fix displaying issues when the custom field list has no options
Basically this has happened a few times when importing a project into
pagure, the custom field of type list is created, given a value but no
data field (ie: no list of options).
This breaks the entire page since we fails at offering the different
options for that field.

With this fix, we do not fix the root cause but we do fix the symptoms.